+++ cloud-core/src/main/java/cloud/commandframework/CommandManager.java
@@ -55,6 +55,7 @@ import cloud.commandframework.meta.CommandMeta;
 import cloud.commandframework.permission.CommandPermission;
 import cloud.commandframework.permission.OrPermission;
 import cloud.commandframework.permission.Permission;
+import cloud.commandframework.permission.PredicatePermission;
 import cloud.commandframework.services.ServicePipeline;
 import cloud.commandframework.services.State;
 import io.leangen.geantyref.TypeToken;
@@ -277,6 +278,7 @@ public abstract class CommandManager<C> {
      * @param permission Permission node
      * @return {@code true} if the sender has the permission, else {@code false}
      */
+    @SuppressWarnings("unchecked")
     public boolean hasPermission(
             final @NonNull C sender,
             final @NonNull CommandPermission permission
@@ -286,6 +288,9 @@ public abstract class CommandManager<C> {
         }
         if (permission instanceof Permission) {
             return hasPermission(sender, permission.toString());
+        }
+        if (permission instanceof PredicatePermission) {
+            return ((PredicatePermission<C>) permission).hasPermission(sender);
         }
         for (final CommandPermission innerPermission : permission.getPermissions()) {
             final boolean hasPermission = this.hasPermission(sender, innerPermission);
@@ -353,11 +358,39 @@ public abstract class CommandManager<C> {
      * @param description Description for the root literal
      * @param meta        Command meta
      * @return Builder instance
+     * @deprecated for removal since 1.4.0. Use {@link #commandBuilder(String, Collection, Description, CommandMeta)} instead.
      */
+    @Deprecated
     public Command.@NonNull Builder<C> commandBuilder(
             final @NonNull String name,
             final @NonNull Collection<String> aliases,
             final @NonNull Description description,
+            final @NonNull CommandMeta meta
+    ) {
+        return commandBuilder(name, aliases, (ArgumentDescription) description, meta);
+    }
+
+    /**
+     * Create a new command builder. This will also register the creating manager in the command
+     * builder using {@link Command.Builder#manager(CommandManager)}, so that the command
+     * builder is associated with the creating manager. This allows for parser inference based on
+     * the type, with the help of the {@link ParserRegistry parser registry}
+     * <p>
+     * This method will not register the command in the manager. To do that, {@link #command(Command.Builder)}
+     * or {@link #command(Command)} has to be invoked with either the {@link Command.Builder} instance, or the constructed
+     * {@link Command command} instance
+     *
+     * @param name        Command name
+     * @param aliases     Command aliases
+     * @param description Description for the root literal
+     * @param meta        Command meta
+     * @return Builder instance
+     * @since 1.4.0
+     */
+    public Command.@NonNull Builder<C> commandBuilder(
+            final @NonNull String name,
+            final @NonNull Collection<String> aliases,
+            final @NonNull ArgumentDescription description,
             final @NonNull CommandMeta meta
     ) {
         return Command.<C>newBuilder(
@@ -393,7 +426,7 @@ public abstract class CommandManager<C> {
         return Command.<C>newBuilder(
                 name,
                 meta,
-                Description.empty(),
+                ArgumentDescription.empty(),
                 aliases.toArray(new String[0])
         ).manager(this);
     }
@@ -413,11 +446,40 @@ public abstract class CommandManager<C> {
      * @param description Description for the root literal
      * @param aliases     Command aliases
      * @return Builder instance
+     * @deprecated for removal since 1.4.0. Use {@link #commandBuilder(String, CommandMeta, ArgumentDescription, String...)}
+     *      instead.
      */
+    @Deprecated
     public Command.@NonNull Builder<C> commandBuilder(
             final @NonNull String name,
             final @NonNull CommandMeta meta,
             final @NonNull Description description,
+            final @NonNull String... aliases
+    ) {
+        return this.commandBuilder(name, meta, (ArgumentDescription) description, aliases);
+    }
+
+    /**
+     * Create a new command builder. This will also register the creating manager in the command
+     * builder using {@link Command.Builder#manager(CommandManager)}, so that the command
+     * builder is associated with the creating manager. This allows for parser inference based on
+     * the type, with the help of the {@link ParserRegistry parser registry}
+     * <p>
+     * This method will not register the command in the manager. To do that, {@link #command(Command.Builder)}
+     * or {@link #command(Command)} has to be invoked with either the {@link Command.Builder} instance, or the constructed
+     * {@link Command command} instance
+     *
+     * @param name        Command name
+     * @param meta        Command meta
+     * @param description Description for the root literal
+     * @param aliases     Command aliases
+     * @return Builder instance
+     * @since 1.4.0
+     */
+    public Command.@NonNull Builder<C> commandBuilder(
+            final @NonNull String name,
+            final @NonNull CommandMeta meta,
+            final @NonNull ArgumentDescription description,
             final @NonNull String... aliases
     ) {
         return Command.<C>newBuilder(
@@ -453,7 +515,7 @@ public abstract class CommandManager<C> {
         return Command.<C>newBuilder(
                 name,
                 meta,
-                Description.empty(),
+                ArgumentDescription.empty(),
                 aliases
         ).manager(this);
     }
@@ -476,10 +538,40 @@ public abstract class CommandManager<C> {
      * @return Builder instance
      * @throws UnsupportedOperationException If the command manager does not support default command meta creation
      * @see #createDefaultCommandMeta() Default command meta creation
+     * @deprecated for removal since 1.4.0. Use {@link #commandBuilder(String, ArgumentDescription, String...)} instead.
      */
+    @Deprecated
     public Command.@NonNull Builder<C> commandBuilder(
             final @NonNull String name,
             final @NonNull Description description,
+            final @NonNull String... aliases
+    ) {
+        return this.commandBuilder(name, (ArgumentDescription) description, aliases);
+    }
+
+    /**
+     * Create a new command builder using default command meta created by {@link #createDefaultCommandMeta()}.
+     * <p>
+     * This will also register the creating manager in the command
+     * builder using {@link Command.Builder#manager(CommandManager)}, so that the command
+     * builder is associated with the creating manager. This allows for parser inference based on
+     * the type, with the help of the {@link ParserRegistry parser registry}
+     * <p>
+     * This method will not register the command in the manager. To do that, {@link #command(Command.Builder)}
+     * or {@link #command(Command)} has to be invoked with either the {@link Command.Builder} instance, or the constructed
+     * {@link Command command} instance
+     *
+     * @param name        Command name
+     * @param description Description for the root literal
+     * @param aliases     Command aliases
+     * @return Builder instance
+     * @throws UnsupportedOperationException If the command manager does not support default command meta creation
+     * @see #createDefaultCommandMeta() Default command meta creation
+     * @since 1.4.0
+     */
+    public Command.@NonNull Builder<C> commandBuilder(
+            final @NonNull String name,
+            final @NonNull ArgumentDescription description,
             final @NonNull String... aliases
     ) {
         return Command.<C>newBuilder(
@@ -516,7 +608,7 @@ public abstract class CommandManager<C> {
         return Command.<C>newBuilder(
                 name,
                 this.createDefaultCommandMeta(),
-                Description.empty(),
+                ArgumentDescription.empty(),
                 aliases
         ).manager(this);
     }